Use Rails Asset Pipeline helpers for asset urls
This makes sure that we're generating the URLs correctly - in production
this means we use the CloudFront CDN url, not our actual app url. Also,
it will use the AssetPipeline to append the correct hash to the image
name, so we can change the image later, but still break any cache the
user's browser has done.