Force UTF-8 encoding on imported beers
They come in as ASCII-8BIT (binary) and then choke when trying to be
converted for search. So, we assume the files are UTF-8 encoded and then
just tell Ruby, "hey, these are UTF-8... trust me!" We need to make sure
instructions for exporting CSV files include getting them UTF-8 encoded.