Add pg_trgm Trigram extension, but don't use
I can't get it to play nicely with the unaccent function. It seems we
can only have one or the other, for now. Likely we need some GiST
indexes on the trigrams, but that's a bit more work that I'm willing to
do right now. http://www.postgresql.org/docs/9.3/static/pgtrgm.html

Also, we could consider using dedicated search, like Elasticsearch, if
we find this PostgreSQL search to not work well enough.