Inject default_url_options for ActiveModel::Serializes
The 0.9.* series now supports url_helpers, but they don't pull in the
default_url_options automatically. Further, their integration with
ActionController does NOT pass along the `url_options` for the current
Request. So we must hack our feature specs to make sure we know the host
and port on which the Capybara web server is running. UGH!